+++ packages/terra-core-docs/src/terra-dev-site/doc/dropdown-button/example/DefaultDropdownButton.jsx
@@ -8,9 +8,7 @@ const Example = () => {
     <React.Fragment>
       <DropdownButton
         label="Export"
-        inputAttrs={{
-          'aria-label': 'default dropdown',
-        }}
+        aria-label="default dropdown"
       >
         <Item label="Export as PDF" onSelect={() => setMessage('Export as PDF clicked')} />
         <Item label="Export as Plain Text" onSelect={() => setMessage('Export as Plain Text clicked')} />