+++ packages/terra-core-docs/src/terra-dev-site/doc/dropdown-button/example/DefaultSplitButton.jsx
@@ -9,9 +9,7 @@ const Example = () => {
       <SplitButton
         primaryOptionLabel="Reply"
         onSelect={() => setMessage('Reply clicked')}
-        inputAttrs={{
-          'aria-label': 'default split',
-        }}
+        aria-label="default split"
       >
         <Item label="Reply All" onSelect={() => setMessage('Reply All clicked')} />
         <Item label="Forward" onSelect={() => setMessage('Forward clicked')} />