+++ packages/terra-dropdown-button/tests/jest/DropdownButton.test.jsx
@@ -96,18 +96,15 @@ describe('Dropdown Button', () => {
   });
 
   it('should set the aria-label property from ./translations', () => {
-    const inputAttrs = {
-      'aria-label': 'Button Aria Label',
-    };
     const wrapper = shallowWithIntl(
       <IntlProvider locale="en" messages={translationsFile}>
-        <DropdownButton label="Primary Option" inputAttrs={inputAttrs} id="dropDown">
+        <DropdownButton label="Primary Option" aria-label="Button Aria Label" id="dropDown">
           <Item label="PDF" onSelect={() => {}} />
         </DropdownButton>
       </IntlProvider>,
     ).dive().dive();
     wrapper.setState({ selectText: 'PDF' });
     const dropdownButtonAriaLabelValue = wrapper.find('#dropDown button').prop('aria-label');
-    expect(dropdownButtonAriaLabelValue).toEqual(`PDF, ${translationsFile['Terra.dropdownButton.selected']}, Primary Option, ${inputAttrs['aria-label']}`);
+    expect(dropdownButtonAriaLabelValue).toEqual(`PDF, ${translationsFile['Terra.dropdownButton.selected']}, Primary Option, Button Aria Label`);
   });
 });