+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -1,11 +1,9 @@
 # Changelog
 
 ## Unreleased
-
-* Updated
-  Added inputAttrs prop for dropdown button.
    
 * Added
+   * Updated aria-label for dropdown button.
    * Added focus styles and screenreader support to announce expand/collapse state and index for dropdown list items.
 
 ## 1.34.0 - (May 9, 2023)